Books feature and matrix feature WIP

+++ school_data_hub_flutter/lib/features/authorizations/domain/authorization_manager.dart
@@ -6,17 +6,20 @@ import 'package:school_data_hub_client/school_data_hub_client.dart';
 import 'package:school_data_hub_flutter/app_utils/custom_encrypter.dart';
 import 'package:school_data_hub_flutter/common/services/notification_service.dart';
 import 'package:school_data_hub_flutter/core/client/client_helper.dart';
-import 'package:school_data_hub_flutter/core/session/serverpod_session_manager.dart';
+import 'package:school_data_hub_flutter/core/session/hub_session_manager.dart';
 import 'package:school_data_hub_flutter/features/authorizations/data/authorization_api_service.dart';
 import 'package:school_data_hub_flutter/features/pupil/domain/models/pupil_proxy.dart';
 import 'package:watch_it/watch_it.dart';
 
-final _notificationService = di<NotificationService>();
-final _authorizationApiService = AuthorizationApiService();
-final _serverpodSessionManager = di<ServerpodSessionManager>();
-final _cacheManager = di<DefaultCacheManager>();
-
 class AuthorizationManager with ChangeNotifier {
+  final _notificationService = di<NotificationService>();
+
+  final _authorizationApiService = AuthorizationApiService();
+
+  final _hubSessionManager = di<HubSessionManager>();
+
+  final _cacheManager = di<DefaultCacheManager>();
+
   ValueListenable<List<Authorization>> get authorizations => _authorizations;
 
   final _authorizations = ValueNotifier<List<Authorization>>([]);
@@ -63,7 +66,9 @@ class AuthorizationManager with ChangeNotifier {
 
   Future<void> fetchAuthorizations() async {
     final authorizations = await _authorizationApiService.fetchAuthorizations();
-
+    if (authorizations == null) {
+      return;
+    }
     _updateAuthsInCollections(authorizations);
     _notificationService.showSnackBar(NotificationType.success,
         '${authorizations.length} Einwilligungen geladen');
@@ -76,10 +81,12 @@ class AuthorizationManager with ChangeNotifier {
     String description,
     List<int> pupilIds,
   ) async {
-    final createdBy = _serverpodSessionManager.userName;
-    final Authorization authorization = await _authorizationApiService
+    final createdBy = _hubSessionManager.userName;
+    final Authorization? authorization = await _authorizationApiService
         .postAuthorizationWithPupils(name, description, createdBy!, pupilIds);
-
+    if (authorization == null) {
+      return;
+    }
     _authorizationsMap[authorization.id!] = authorization;
     _authorizations.value = _authorizationsMap.values.toList();
 
@@ -100,7 +107,9 @@ class AuthorizationManager with ChangeNotifier {
     final updatedAuth = await ClientHelper.apiCall(
         call: () => _authorizationApiService.updateAuthorization(
             authId, name, description, membersToUpdate));
-
+    if (updatedAuth == null) {
+      return;
+    }
     _updateAuthsInCollections([updatedAuth]);
     _notificationService.showSnackBar(
         NotificationType.success, 'Einwilligung geändert');
@@ -110,9 +119,7 @@ class AuthorizationManager with ChangeNotifier {
 
   Future<void> deleteAuthorization(int authId) async {
     final confirm = await _authorizationApiService.deleteAuthorization(authId);
-    if (!confirm) {
-      _notificationService.showSnackBar(
-          NotificationType.error, 'Einwilligung konnte nicht gelöscht werden');
+    if (confirm == null) {
       return;
     }
     _authorizationsMap.remove(authId);
@@ -137,6 +144,9 @@ class AuthorizationManager with ChangeNotifier {
     );
     final updatedPupilAuth = await _authorizationApiService
         .updatePupilAuthorization(pupilAuthUpdate);
+    if (updatedPupilAuth == null) {
+      return;
+    }
     _updatePupilAuthInCollections(updatedPupilAuth);
 
     _notificationService.showSnackBar(
@@ -150,9 +160,12 @@ class AuthorizationManager with ChangeNotifier {
     int pupilAuthId,
   ) async {
     final encryptedFile = await customEncrypter.encryptFile(file);
-    final createdBy = _serverpodSessionManager.userName;
+    final createdBy = _hubSessionManager.userName;
     final pupilAuth = await _authorizationApiService
         .addFileToPupilAuthorization(pupilAuthId, encryptedFile, createdBy!);
+    if (pupilAuth == null) {
+      return;
+    }
     _updatePupilAuthInCollections(pupilAuth);
 
     return;
@@ -164,6 +177,9 @@ class AuthorizationManager with ChangeNotifier {
   ) async {
     final pupilAuth =
         await _authorizationApiService.removeFileFromPupilAuthorization(authId);
+    if (pupilAuth == null) {
+      return;
+    }
     _updatePupilAuthInCollections(pupilAuth);
     _cacheManager.removeFile(cacheKey);
